Replaced raw loop with STL algorithm
I followed Sean Parent's advice, but at what cost?

For real though, I like that this is possible but it feels like an
excessively unreadable solution. The raw loop I replaced was already
pretty simple and readable. It's not as if I need the potential
algorithm advantanges of an STL algorithm since this function won't
even see 10 elements in the lifetime of the program.