Remove include directory
These are not public headers. The headers mainly exist to allow
more source code to live in separate .cpp files with a nice side
effect that it's easier to test as well. The existence of an include
directory implies you might ingest this project and use them which
is certainly not intended.